refactor: move tests related to denials
Although these tests are also part of the corrections, I think they make
more sense in the denials test class. They revert changes made by the
corrections, but that reverting is part of the denied.
